# docs/borga-api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BORGA",
    "description": "BORGA (BOaRd Games Application) provides access, through a web interface (hypermedia), to some of the features provided by the [Board Game Atlas](https://www.boardgameatlas.com/) website, making use of its [Web API](https://www.boardgameatlas.com/api/docs) for this purpose. For all group operations, a user token must be sent in the Authorization header using a Bearer Token. \n Authors: André Jesus, Nyckollas Brandão and André Santos.",
    "termsOfService": "http://swagger.io/terms/",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/games/search": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Searches games by name.",
        "parameters": [
          {
            "name": "gameName",
            "in": "query",
            "description": "The name of the game to search.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limits the number of results returned. The max limit is 100. The default is 30.",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "description": "Provide the name of another parameter and it will order the results according to it.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ascending",
            "in": "query",
            "description": "This determines which direction the list is shown in based on the order_by parameter.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "example": {
                  "games": [
                    {
                      "id": "OIXt3DmJU0",
                      "name": "Catan",
                      "description": "\nThe women and men of your expedition build the first two settlements. Fortunately, the land is rich in natural resources. You build roads and new settlements that eventually become cities. Will you succeed in gaining supremacy on Catan? Barter trade dominates the scene. Some resources you have in abundance, other resources are scarce. Ore for wool, brick for lumber - you trade according to what is needed for your current building projects. Proceed strategically! If you found your settlements in the right places and skillfully trade your resources, then the odds will be in your favor. But your opponents are smart too.\n\r\n\nTo begin the game, we build the game board using hexagonal terrain tiles. Catan is born - a beautiful island with mountains, pastures, hills, fields, and forests, surrounded by the sea.\n\r\n\nEach of us places two small houses on spaces where three terrain hexes meet. They are our starting settlements.\n\r\n\nAnd so it begins. I roll two dice. An “11”! Each terrain hex is marked with a die roll number. Each player who owns a settlement adjacent to a terrain hex marked with the number rolled receives a resource produced by this hex. Hills produce brick, forests produce lumber, mountains produce ore, fields produce grain, and pastures produce wool.\n\r\n\nWe use these resources to expand across Catan: we build roads and new settlements, or we upgrade our existing settlements to cities. For example, a road costs 1 brick and 1 lumber. If we do not have the necessary resources, we can acquire them by trading with our opponents.\n\r\n\nEach settlement is worth 1 victory point and each city is worth 2 victory points. If you expand cleverly, you may be the first player to reach 10 victory points and thus win the game!\n",
                      "url": "https://www.boardgameatlas.com/game/OIXt3DmJU0/catan",
                      "image_url": "https://s3-us-west-1.amazonaws.com/5cc.images/games/uploaded/1629324722072.jpg",
                      "publisher": "KOSMOS",
                      "amazon_rank": 133,
                      "price": "32.97",
                      "mechanics": [
                        "Dice Rolling",
                        "Network and Route Building",
                        "Trading"
                      ],
                      "categories": [
                        "Dice",
                        "Economic",
                        "Family Game",
                        "Negotiation"
                      ]
                    },
                    {
                      "id": "EvYzPCnCGG",
                      "name": "Catan: 5-6 Player Extension",
                      "description": "Now five to six players can explore and settle Catan! In the Catan: 5-6 Player Extension you control a group of settlers exploring and taming the uncharted lands of Catan. Embark on a quest to settle the rich island, competing against more opponents for added fun. ● Add 1-2 more friends or family without sacrificing ease of play. ● Add green and brown settlers and expand your island with 2 more harbors and 11 unique, new terrain tiles. This is NOT a complete game! It is only a game extension for 5-6 players. You need a Catan game (aka The Settlers of Catan) in order to play with this extension! Components ● 11 Hexagonal Regional Tiles ● 2 Frame Pieces with Harbors ● 2 All-Sea Frame Pieces ● 10 Wooden Settlements ● 8 Wooden Cities ● 30 Wooden Roads ● 25 Resource Cards ● 9 Development Cards ● 28 Number Tokens",
                      "url": "https://www.boardgameatlas.com/game/EvYzPCnCGG/catan-56-player-extension",
                      "image_url": "https://s3-us-west-1.amazonaws.com/5cc.images/games/uploaded/1559254797611-51uLrWkNt0L.jpg",
                      "publisher": "Catan Studio",
                      "amazon_rank": 534,
                      "price": "22.50",
                      "mechanics": [
                        "Dice Rolling",
                        "Hand Management",
                        "Memory",
                        "Modular Board",
                        "Trading",
                        "Network and Route Building"
                      ],
                      "categories": [
                        "Expansion",
                        "Negotiation"
                      ]
                    },
                    {
                      "id": "NoWy3nPYfj",
                      "name": "Catan: Seafarers Expansion",
                      "description": "\nIn Catan: Seafarers you control a group of bold seafaring settlers exploring and taming the wild, uncharted Isles of Catan. Each game brings you to new seas and new lands. You might even discover the precious gold fields (a very valuable new terrain).<br /><br />Every game is unique and full of fun!<br /><br />Embark on an epic quest to settle the home island, build ships, and chart nearby waters. Guide your brave seafarers to victory through cunning trades and wise development. Be first to create new sea lanes and settle the newly-discovered isles.<br /><br />The best strategy and a nice dash of luck decide who will be the undisputed ruler of the rich Isles of Catan!<br />Capture the adventure of seafaring, exploration, and trade.<br />Catan: Seafarers is a 3-4 player expansion for the best-selling Catan board game. Share some fun at sea!<br /><br />Components<br />60 Wooden Ships<br />1 Pirate Ship<br />6 Frame Pieces<br />19 Sea Hexes<br />11 Hexagonal Region Tiles<br />50 Catan Chits<br />10 Number Tokens<br />10 Harbor Tokens<br />1 Rulebook and Scenarios\n",
                      "url": "https://www.boardgameatlas.com/game/NoWy3nPYfj/catan-seafarers-expansion",
                      "image_url": "https://s3-us-west-1.amazonaws.com/5cc.images/games/uploaded/1559257503054-51u0777bL3L.jpg",
                      "publisher": "Catan Studio",
                      "amazon_rank": 2618,
                      "price": "29.99",
                      "mechanics": [],
                      "categories": [
                        "Expansion"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "example": {
                  "cause": {
                    "code": 1001,
                    "name": "NOT_FOUND",
                    "message": "The item does not exist {\"gameName\":\"Catan\"}",
                    "info": {
                      "gameName": "Catan"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/user/{userId}/groups": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Creates a group providing its name and description.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Group information",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "groupName": {
                    "type": "string"
                  },
                  "groupDescription": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "example": {
                  "Created group": {
                    "id": "574d6214-037b-4de0-a773-77a33e23b4ad",
                    "name": "Negotiation games",
                    "description": "This is a group of negotiation games."
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/json": {
                "example": {
                  "cause": {
                    "code": 1005,
                    "name": "UNAUTHENTICATED",
                    "message": "Invalid or missing token",
                    "info": "Please insert a valid user token"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Lists all groups.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "example": {
                  "45103531-8450-4970-8ebd-8b3e5d316382": {
                    "name": "NegG",
                    "description": "Negotiation games"
                  },
                  "f670933b-722b-43cb-97d1-aa89f3336085": {
                    "name": "Negotiation and Economic Games",
                    "description": "This is a group of negotiation and economic games."
                  },
                  "839c0857-6e6d-4c6d-bef0-ac9d52e45819": {
                    "name": "Role-Playing Games",
                    "description": "This is a group of role-playing games."
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/json": {
                "example": {
                  "cause": {
                    "code": 1005,
                    "name": "UNAUTHENTICATED",
                    "message": "Invalid or missing token",
                    "info": "Please insert a valid user token"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/groups/{groupId}": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Edits a group by changing its name and description.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "description": "The group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Group information",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newGroupName": {
                    "type": "string"
                  },
                  "newGroupDescription": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "example": {
                  "Edited group": {
                    "id": "f670933b-722b-43cb-97d1-aa89f3336085",
                    "name": "Negotiation and Economic Games",
                    "description": "This is a group of negotiation and economic games."
                  }
                }
              }
            }
          },
          "404": {
            "description": "group doesn't exist",
            "content": {
              "application/json": {
                "example": {
                  "cause": {
                    "code": 1002,
                    "name": "NOT_FOUND",
                    "message": "The item does not exist",
                    "info": {
                      "groupId": "572bc968-87d1-436d-b455-26acf7f58fdc"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Deletes a group.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "description": "The group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "example": {
                  "Deleted group": {
                    "id": "45103531-8450-4970-8ebd-8b3e5d316382",
                    "name": "NegG",
                    "description": "Negotiation games"
                  }
                }
              }
            }
          },
          "404": {
            "description": "group doesn't exist",
            "content": {
              "application/json": {
                "example": {
                  "cause": {
                    "code": 1002,
                    "name": "NOT_FOUND",
                    "message": "The item does not exist",
                    "info": {
                      "groupId": "28248a5a-8bd1-422b-87e8-1b420fe7e1ee"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Gets the details of a group.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "description": "The group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "example": {
                  "id": "839c0857-6e6d-4c6d-bef0-ac9d52e45819",
                  "name": "Role-Playing Games",
                  "description": "This is a group of role-playing games.",
                  "games": {
                    "NoWy3nPYfj": "Catan: Seafarers Expansion",
                    "OIXt3DmJU0": "Catan"
                  }
                }
              }
            }
          },
          "404": {
            "description": "user or group not found",
            "content": {
              "application/json": {
                "example": {
                  "cause": {
                    "code": 1002,
                    "name": "NOT_FOUND",
                    "message": "The item does not exist",
                    "info": {
                      "groupId": "ff5739dc-e856-46af-be1f-1ea1fc306382"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/groups/{groupId}/games": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Adds a game to a group.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "description": "The group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Game information",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gameId": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "example": {
                  "Added game": {
                    "id": "OIXt3DmJU0",
                    "name": "Catan",
                    "description": "\nThe women and men of your expedition build the first two settlements. Fortunately, the land is rich in natural resources. You build roads and new settlements that eventually become cities. Will you succeed in gaining supremacy on Catan? Barter trade dominates the scene. Some resources you have in abundance, other resources are scarce. Ore for wool, brick for lumber - you trade according to what is needed for your current building projects. Proceed strategically! If you found your settlements in the right places and skillfully trade your resources, then the odds will be in your favor. But your opponents are smart too.\n\r\n\nTo begin the game, we build the game board using hexagonal terrain tiles. Catan is born - a beautiful island with mountains, pastures, hills, fields, and forests, surrounded by the sea.\n\r\n\nEach of us places two small houses on spaces where three terrain hexes meet. They are our starting settlements.\n\r\n\nAnd so it begins. I roll two dice. An “11”! Each terrain hex is marked with a die roll number. Each player who owns a settlement adjacent to a terrain hex marked with the number rolled receives a resource produced by this hex. Hills produce brick, forests produce lumber, mountains produce ore, fields produce grain, and pastures produce wool.\n\r\n\nWe use these resources to expand across Catan: we build roads and new settlements, or we upgrade our existing settlements to cities. For example, a road costs 1 brick and 1 lumber. If we do not have the necessary resources, we can acquire them by trading with our opponents.\n\r\n\nEach settlement is worth 1 victory point and each city is worth 2 victory points. If you expand cleverly, you may be the first player to reach 10 victory points and thus win the game!\n",
                    "url": "https://www.boardgameatlas.com/game/OIXt3DmJU0/catan",
                    "image_url": "https://s3-us-west-1.amazonaws.com/5cc.images/games/uploaded/1629324722072.jpg",
                    "amazon_rank": 133,
                    "price": "29.99",
                    "mechanics": [
                      "Dice Rolling",
                      "Network and Route Building",
                      "Trading"
                    ],
                    "categories": [
                      "Dice",
                      "Economic",
                      "Family Game",
                      "Negotiation"
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "group doesn't exist",
            "content": {
              "application/json": {
                "example": {
                  "cause": {
                    "code": 1002,
                    "name": "NOT_FOUND",
                    "message": "The item does not exist",
                    "info": {
                      "groupId": "ff5739dc-e856-46af-be1f-1ea1fc306382"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}